backend/requirements.txt
Flask
earthengine-api
flask-cors
google-generativeai
python-dotenv